refactor: 이메일 인증 관련 컴포넌트 재사용
