+++ src/pages/auth/register/verify.tsx
@@ -1,31 +1,13 @@
-import Link from 'next/link';
-import { useRouter } from 'next/router';
 import styled from '@emotion/styled';
-import ByEmail from '@/components/auth/verify/ByEmail';
-import Logo from '@/components/common/Logo';
-import Responsive from '@/components/common/Responsive';
+import VerifyForm from '@/components/auth/verify/VerifyForm';
 import FullHeightLayout from '@/components/layout/FullHeightLayout';
-import { tilyLinks } from '@/constants/links';
 import { setLayout } from '@/utils/layout';
 
-export const RegisterVerifyPage = () => {
-  const router = useRouter();
-
+const RegisterVerifyPage = () => {
   return (
-    <StyledVerifyPage>
-      <Responsive device="desktop">
-        <button onClick={() => router.push(tilyLinks.login())}>
-          <Logo />
-        </button>
-      </Responsive>
-      <Responsive device="mobile">
-        <button onClick={() => router.push(tilyLinks.login())}>
-          <Logo imageSize={42} />
-        </button>
-      </Responsive>
-      <ByEmail type="register" />
-      <StyledLoginButton href={tilyLinks.login()}>이미 계정이 있나요? 로그인하기</StyledLoginButton>
-    </StyledVerifyPage>
+    <VerifyPage>
+      <VerifyForm location="register" />
+    </VerifyPage>
   );
 };
 
@@ -33,7 +15,7 @@ setLayout(RegisterVerifyPage, FullHeightLayout);
 
 export default RegisterVerifyPage;
 
-const StyledVerifyPage = styled.div`
+export const VerifyPage = styled.main`
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -45,20 +27,5 @@ const StyledVerifyPage = styled.div`
 
   @media ${({ theme }) => theme.mediaQuery.sm} {
     height: 100dvh;
-  }
-`;
-
-const StyledLoginButton = styled(Link)`
-  align-self: flex-end;
-  margin-top: 0.5rem;
-  padding: 0;
-  font-size: 0.8rem;
-  font-weight: 600;
-  color: ${({ theme }) => theme.colors.gray_800};
-
-  transition: all 0.1s;
-
-  &:hover {
-    color: #5d5d5d;
   }
 `;
